% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segtree.R
\name{performance}
\alias{performance}
\alias{performance.segtree}
\alias{print.performance.segtree}
\title{Evaluate the Performance of a Segmentation}
\usage{
performance(x, ...)

\method{performance}{segtree}(tree, ...)

\method{print}{performance.segtree}(tree, details = c(1, 2, 3, 0), ...)
}
\arguments{
\item{tree}{An object of class \code{segtree}}

\item{details}{The level of details to show in the tree structure (see
\code{\link{structure.segtree}})}
}
\value{
An object of class \code{performance.segtree} with the most important
  stats of the tree's performance
}
\description{
Compares the Gini Index before and after applying the segmentation.
}

