% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segtree.R
\name{segtree}
\alias{print.segtree}
\alias{segtree}
\title{Create Segmentation Trees}
\usage{
segtree(formula, data, segvars, fast = FALSE, ...)

\method{print}{segtree}(tree, ...)
}
\arguments{
\item{formula}{A formula specifying the response and features for the logistic regressions}

\item{data}{A \code{data.frame} containing the variables specified in formula and in \code{segvars}}

\item{segvars}{A character vector with the names of the segmentation variables to be tried}

\item{fast}{If \code{FALSE} (the default), evaluate the splits for each node. Otherwise construct a simpler tree which takes less time to build}
}
\value{
An object of class \code{segtree}
}
\description{
\code{segtree} builds a basic segmentation tree whose nodes can then be splitted with \code{fork} and evaluated with a variety of functions.
}
\seealso{
\code{\link{fork}}, \code{\link{summary.segtree}}, \code{\link{performance.segtree}}, \code{\link{split_data}}, \code{\link{split.segtree}}
}

