% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segtree.R
\name{split.segtree}
\alias{print.segtree.split}
\alias{split.segtree}
\title{Evaluate Splits on a Tree}
\usage{
\method{split}{segtree}(tree, leaf)

\method{print}{segtree.split}(s, global_pop = NULL)
}
\arguments{
\item{tree}{An object of class \code{segtree}}

\item{leaf}{Either the name of a node or an actual node of \code{tree}}
}
\value{
A simple object of class \code{segtree.split} containing the results
  of splitting with each remaining segmentation variable
}
\description{
This function is used internally to try out all the remaining segmentation
variables on a given node.
}

