% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{leaf_index}
\alias{leaf_index}
\alias{split_data}
\title{Apply Segmentation to Data}
\usage{
leaf_index(leaf, data)

split_data(tree, index.only = T, newdata = NULL, ...)
}
\arguments{
\item{leaf}{An object of class \code{leaf}}

\item{data}{A \code{data.frame} to be segmented}

\item{tree}{tree An object of class \code{segtree}}

\item{index.only}{Whether to return the segmentation variable (if
\code{TRUE}, the default) or to actually split the data}

\item{newdata}{A \code{data.frame} to be split. If \code{NULL}, then the
tree's data is used}
}
\description{
Returns a factor with the terminal node each observation corresponds to.
Alternatively, returns a list of \code{data.frame}s corresponding to each
terminal node. This function is similar to some predict methods.
}
\details{
The function \code{split_data} a convenient wrapper of
  \code{leaf_index}. \code{leaf_index} simply returns a boolean vector of
  whether each observation is in the given leaf or not.
}

