% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{structure.segtree}
\alias{print.structure.segtree}
\alias{structure.segtree}
\title{Structure of a Segmentation Tree}
\usage{
structure.segtree(tree, leaf_name = "root")

\method{print}{structure.segtree}(s, details = c(0, 1, 2, 3), prefix = "",
  level = 0)
}
\arguments{
\item{tree}{tree An object of class \code{segtree}}

\item{leaf_name}{The name of a leaf in the tree}

\item{details}{The level of details to show in the tree structure. 0 means no
details, 1 means only on terminal nodes, 2 means on all nodes and >= 3
means on all nodes plus extra information}
}
\value{
An object of class \code{structure.segtree}, which draws the tree's
  structure in a text graph
}
\description{
Shows a tree-like data structure with the dependences between the nodes of a
\code{segtree}.
}

