% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_segtree.R
\name{summary.segtree}
\alias{print.summary.segtree}
\alias{summary.segtree}
\title{Summarize Segmentation Trees}
\usage{
\method{summary}{segtree}(tree)

\method{print}{summary.segtree}(s, details = c(1, 2, 3, 0))
}
\arguments{
\item{tree}{tree An object of class \code{segtree}}

\item{s}{An object of class \code{summary.segtree}}

\item{details}{The level of details to show in the tree structure (see
\code{\link{structure.segtree}})}
}
\value{
An object of class \code{summary.segtree} which prints the most
  important features and structure of the tree in a nice and easy-to-see way
}
\description{
Show the most important facts about a tree, including a text drawing of its
structure.
}

